projet fonctionnel

+++ main/main.c
@@ -6,17 +6,14 @@
 /*   By: snaggara <snaggara@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/30 10:28:18 by snaggara          #+#    #+#             */
-/*   Updated: 2023/06/02 14:33:13 by snaggara         ###   ########.fr       */
+/*   Updated: 2023/06/05 11:11:30 by snaggara         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "pipex.h"
 
 
-// Samir il te reste a voir les fds open, c'est chelou, tu les vois avec  lsof | grep sam-acer | grep /home/sam-acer/Documents/42/pipex
-// en ligne de commande.
-// Apparemment, il faut fermer tout les fds à la fin d'un programme, ne pas en laisser un seul d'ouvert
-// Meme pas stdin (chelou mais bon)
+
 int	main(int ac, char **av, char **envp)
 {
 	t_data	*d;
@@ -24,10 +21,33 @@ int	main(int ac, char **av, char **envp)
 	d = ft_init(ac, av, envp);
 	if (!d)
 		return (0);
+	// if (!ft_create_stdout(d))
+	// 	return (0);
+	
+	if (!ft_test_cmds(d))
+		return (0);
 	ft_fork(d);
 	ft_free_double_tab(d->path);
 	if (d->here_doc)
 		unlink(TMP_FILE_NAME);
 	free(d);
 	return (0);
+}
+
+int	ft_test_cmds(t_data *d)
+{
+	int	status;
+
+	status = 1;
+	if (!ft_cmd_valid(d, d->cmd1))
+		status = 0;
+	if (!ft_cmd_valid(d, d->cmd2))
+		status = 0;
+	if (!status)
+	{
+		ft_close_all_fds();
+		ft_free_double_tab(d->path);
+		free(d);
+	}
+	return (status);
 }
